add delete confirmation and splash screen

+++ lib/widgets/posts.dart
@@ -228,6 +228,27 @@ class _PostsState extends State<Posts> {
   }
 
   void _deleteItem(String id, String URL) async {
+    final result = await showDialog<bool>(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Are you sure?'),
+        content: const Text('This action will permanently delete this post'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context, false),
+            child: const Text('Cancel'),
+          ),
+          TextButton(
+            onPressed: () => Navigator.pop(context, true),
+            child: const Text('Delete'),
+          ),
+        ],
+      ),
+    );
+
+    if (result == null || !result) {
+      return;
+    }
     FirebaseStorage.instance.refFromURL(URL).delete();
     await _firestore.collection('posts').doc(id).delete();
   }
